Refactor shallow schema (#1350)
* `shallow_schema` refactor to make its generation more maintainable.

Overview
========

This PR introduces a kind of _mixin field_ `ModelValidation` with model type validation.
All the Marshmallow fields are imported into `qiskit.validation.fields` and extended to
include the mixin. The same happens with polymorphic fields and containers.

Validation is possible thanks to a special copy of the model schema, the _shallow schema_,
which replaces the `_deserialize` private method of its fields (at instance-level) to use
the new `validate_model` from the `ModelValidator` mixin.

Validating models consist into checking that the values are of the corresponding types.
Specifically:

1. For regular fields, the type they deserialize to (deserialization type).
2. For fields containers (such as lists), the deserialization type of the contained field.
3. For schema containers (such as nested schemas), the type of the schema's model.
4. For polymorphic fields, one of the deserialization types of the choices.
5. For polymorphic schemas, one of the types of the possible schemas' model.

While validating, nothing more than the type of the value is considered. Validation
assumes that the values of the inner properties are already valid or they would have
failed during their instantiation. For instance, consider:

```python
Book(title="The Terror", author=Person(name="Dan Simmons"))
```

`Person` instantiation is executed first. If successful, validating `Book` does not
require `author`'s internal structure (i.e. `name`) to be validated again since otherwise,
construction of `Person` would have failed.
---

This refactor continues patching the method `_deserialize` of the schema fields but,
instead of deciding how to patch it according to the type, it makes `_deserialize`
to call a new custom method `_validate_model` with the desired behaviour.

The different strategies from the if/elif/else block of the original implementation are
now spread over the fields defined in qiskit.validation.fields

This package also contains versions of marshmallow.fields that include their own
`_validate_model` method.

* Organize the fields into several modules and factor out validation.

This PR split the fields into general fields, containers and polymorphic. Each module
has the definition of the different kind of fields.

It also provides a base field for those able of model type validation and a general
mechanism to perform this validation.

* Polishing docstrings and reviewing container and polymorphic validations.

* Rename to make intentions clear and fixed fields hierarchy

* Moving tests to its own folder

* Split tests into different categories

* Docstrings fixes and clarifications.

* Style, convention and doc tweaks

* Update CHANGELOG
