+++ CHANGELOG.rst
@@ -66,8 +66,8 @@ Changed
 - `IBMQ.save_account()` now takes an `overwrite` option to replace an existing
   account on disk. Default is False (#1295).
 - Backend and Provider methods defined in the specification use model objects
-  rather than dicts, along with validation against schemas (#1249, #1277). The
-  updated methods include:
+  rather than dicts, along with validation against schemas (#1249, #1277,
+  #1350). The updated methods include:
     - ``backend.status()``(#1301).
     - ``backend.configuration()`` (and ``__init__``) (#1323).
     - ``backend.properties()``, returning ``None`` for sims (#1331, #1401).
