+++ qiskit/backends/models/backendconfiguration.py
@@ -7,10 +7,10 @@
 
 """Model and schema for backend configuration."""
 
-from marshmallow.fields import Boolean, DateTime, Integer, List, Nested, String
 from marshmallow.validate import Equal, Length, OneOf, Range, Regexp
 
 from qiskit.validation import BaseModel, BaseSchema, bind_schema
+from qiskit.validation.fields import Boolean, DateTime, Integer, List, Nested, String
 
 
 class GateConfigSchema(BaseSchema):
