+++ qiskit/backends/models/backendstatus.py
@@ -7,10 +7,10 @@
 
 """Model and schema for backend status."""
 
-from marshmallow.fields import Boolean, Integer, String
 from marshmallow.validate import Range, Regexp
 
 from qiskit.validation import BaseModel, BaseSchema, bind_schema
+from qiskit.validation.fields import Boolean, Integer, String
 
 
 class BackendStatusSchema(BaseSchema):
