+++ qiskit/validation/__init__.py
@@ -6,4 +6,7 @@
 # the LICENSE.txt file in the root directory of this source tree.
 
 """Models and schemas for Terra."""
-from .base import BaseModel, BaseSchema, bind_schema
+
+from marshmallow import ValidationError
+
+from .base import BaseModel, BaseSchema, bind_schema, ModelTypeValidator
